foodsnap_app/requirements.txt
kivy>=2.1.0
kivymd>=1.1.1
tflite-runtime>=2.9.0
requests>=2.28.0
pillow>=9.0.0
numpy>=1.21.0
plyer>=2.0.0
